+++ p3/synchro.h
@@ -21,7 +21,7 @@ void mutex_lock(struct mutex_t* m);
 void mutex_unlock(struct mutex_t* m);
 
 typedef struct semaphore_t {
-   uint16_t keys;
+   uint8_t keys;
    queue_t queue;
 } semaphore_t;
